Initial commit
This commit adds the current version of code that I use at home. I have
done some work to try and package this, but it is very much a work in
progress and has not been tested anywhere outside of my home. The current
code also includes some hacks that I have had to make to deal with our
prepaid meter tripping when it receives any reverse load - which is
problematic, as the way ESS works there will always be a bit of reverse
load as a large AC load (kettle, geyser etc) turns off. Unless you're
pulling a reasonable amount from the grid, in which case that gives some
buffer for the inverter to compensate and draw less without feeding back.